Added word2vec vocabulary generation
