+++ datasets.py
@@ -9,8 +9,7 @@ from pyannote.audio.embedding.generators import SpeechSegmentGenerator
 from pyannote.database import get_protocol
 from pyannote.database import FileFinder
 from os.path import join
-from tqdm import tqdm
-from collections import deque
+import sts_utils as sts
 
 
 class SimDataset:
@@ -101,97 +100,6 @@ class VoxCeleb1(SimDataset):
         return VoxCelebPartition(self.dev_gen)
 
 
-class SemEvalSegment:
-
-    @staticmethod
-    def find_cluster(clusters, sentence):
-        for i, cluster in enumerate(clusters):
-            if sentence in cluster:
-                return i
-        return None
-
-    def __init__(self, sents):
-        self.sents = sents
-
-    def clusters(self, other_segment, scores, threshold=2.5):
-        """
-        Consider scores as edge weights in a graph of sentences.
-        Search for positive and negative pairs Breadth First Search
-        """
-        clusters = []
-        for i, s in tqdm(enumerate(set(self.sents))):
-            if SemEvalSegment.find_cluster(clusters, s) is not None:
-                continue
-            c = [s]
-            added = {(i, self)}
-            stack = deque()
-            for j, x in enumerate(self.sents):
-                if j != i and x == s:
-                    stack.append((j, other_segment, self, True))
-                    added.add((j, other_segment))
-            while stack:  # is not empty
-                # Retrieve next sentence from the stack
-                j, seg, other_seg, equals_last = stack.popleft()
-                other_sent = seg.sents[j]
-                # Create the pair
-                equals_this = False
-                if scores[j] >= threshold:
-                    if equals_last:
-                        # A = B = C --> A = C (We're putting these 2 in the same cluster)
-                        c.append(other_sent)
-                        equals_this = True
-                # Add dependencies
-                for k, x in enumerate(seg.sents):
-                    if k != j and (k, other_seg) not in added and x == other_sent:
-                        stack.append((k, other_seg, seg, equals_this))
-                        added.add((k, other_seg))
-            if len(c) > 1:
-                clusters.append(c)
-        return clusters
-
-    def pos_neg_pairs(self, other_segment, scores, threshold=2.5):
-        """
-        Consider scores as edge weights in a graph of sentences.
-        Search for positive and negative pairs Breadth First Search
-        """
-        if isinstance(threshold, tuple):
-            tlow, thigh = threshold
-        else:
-            tlow = threshold
-            thigh = threshold
-        pos, neg = [], []
-        for i, s in tqdm(enumerate(set(self.sents))):
-            added = {(i, self)}
-            stack = deque()
-            for j, x in enumerate(self.sents):
-                if j != i and x == s:
-                    stack.append((j, other_segment, self, True))
-                    added.add((j, other_segment))
-            while stack:  # is not empty
-                # Retrieve next sentence from the stack
-                j, seg, other_seg, equals_last = stack.popleft()
-                other_sent = seg.sents[j]
-                # Create the pair
-                equals_this = False
-                if scores[j] >= thigh:
-                    if equals_last:
-                        # A = B = C --> A = C
-                        pos.append((s, other_sent))
-                        equals_this = True
-                    else:
-                        # A != B and B = C --> A != C
-                        neg.append((s, other_sent))
-                elif scores[j] <= tlow and equals_last:
-                    # A = B and B != C --> A != C
-                    neg.append((s, other_sent))
-                # Add dependencies
-                for k, x in enumerate(seg.sents):
-                    if k != j and (k, other_seg) not in added and x == other_sent:
-                        stack.append((k, other_seg, seg, equals_this))
-                        added.add((k, other_seg))
-        return pos, neg
-
-
 class SemEvalPartition(SimDatasetPartition):
 
     def __init__(self, data):
@@ -213,46 +121,6 @@ class SemEvalPartition(SimDatasetPartition):
 class SemEval(SimDataset):
 
     @staticmethod
-    def unique_pairs(xs, ys, scores):
-        seen = set()
-        xunique, yunique, sunique = [], [], []
-        for x, y, score in zip(xs, ys, scores):
-            if (x, y) not in seen:
-                seen.add((x, y))
-                seen.add((y, x))
-                xunique.append(x)
-                yunique.append(y)
-                sunique.append(score)
-        return xunique, yunique, sunique
-
-    @staticmethod
-    def anchor_related_sents(anchor, pairs):
-        anchor_pairs = [(x, y) for x, y in pairs if x == anchor or y == anchor]
-        related = []
-        for pos1, pos2 in anchor_pairs:
-            if pos1 != anchor:
-                related.append(pos1)
-            else:
-                related.append(pos2)
-        return related
-
-    @staticmethod
-    def triplets(unique_sents, pos_pairs, neg_pairs):
-        anchors, positives, negatives = [], [], []
-        for anchor in tqdm(unique_sents):
-            for positive in SemEval.anchor_related_sents(anchor, pos_pairs):
-                for negative in SemEval.anchor_related_sents(anchor, neg_pairs):
-                    anchors.append(anchor)
-                    positives.append(positive)
-                    negatives.append(negative)
-        return anchors, positives, negatives
-
-    @staticmethod
-    def pairs(segment_a, segment_b, scores, threshold=(2, 3)):
-        pos, neg = segment_a.pos_neg_pairs(segment_b, scores, threshold=threshold)
-        return set(pos), set(neg)
-
-    @staticmethod
     def load_partition(partition_path, mode, threshold):
         with open(join(partition_path, 'a.toks'), 'r') as file_a,\
                 open(join(partition_path, 'b.toks'), 'r') as file_b,\
@@ -260,16 +128,16 @@ class SemEval(SimDataset):
             sents_a = [line.strip() for line in file_a.readlines()]
             sents_b = [line.strip() for line in file_b.readlines()]
             scores = [float(line.strip()) for line in score_file.readlines()]
-            sents_a, sents_b, scores = SemEval.unique_pairs(sents_a, sents_b, scores)
+            sents_a, sents_b, scores = sts.unique_pairs(sents_a, sents_b, scores)
             unique_sents = set(sents_a + sents_b)
-            segment_a = SemEvalSegment(sents_a)
-            segment_b = SemEvalSegment(sents_b)
+            segment_a = sts.SemEvalSegment(sents_a)
+            segment_b = sts.SemEvalSegment(sents_b)
             if mode == 'pairs':
-                pos, neg = SemEval.pairs(segment_a, segment_b, scores, threshold)
+                pos, neg = sts.pairs(segment_a, segment_b, scores, threshold)
                 data = np.array([(s1, s2, 0) for s1, s2 in pos] + [(s1, s2, 1) for s1, s2 in neg])
             elif mode == 'triplets':
-                pos, neg = SemEval.pairs(segment_a, segment_b, scores, threshold)
-                anchors, positives, negatives = SemEval.triplets(unique_sents, pos, neg)
+                pos, neg = sts.pairs(segment_a, segment_b, scores, threshold)
+                anchors, positives, negatives = sts.triplets(unique_sents, pos, neg)
                 data = np.array(list(zip(anchors, positives, negatives)))
             elif mode == 'auto':
                 clusters = segment_a.clusters(segment_b, scores, threshold)
@@ -282,17 +150,15 @@ class SemEval(SimDataset):
             np.random.shuffle(data)
         return data
 
-    def __init__(self, path, batch_size, mode='auto', threshold=2.5):
+    def __init__(self, path, vector_path, vocab_path, batch_size, mode='auto', threshold=2.5):
         self.batch_size = batch_size
+        self.vocab, n_inv, n_oov = sts.vectorized_vocabulary(vocab_path, vector_path)
+        print(f"Created vocabulary. {n_inv} words out of {n_inv + n_oov} ({100 * n_inv / (n_inv + n_oov):d}%)")
         self.train_sents = self.load_partition(join(path, 'train'), mode, threshold)
-        # TODO uncomment this when dev partition is available
-        # self.dev_sents = self.load_partition(join(path, 'dev'), mode, threshold)
+        self.dev_sents = self.load_partition(join(path, 'dev'), mode, threshold)
 
     def training_partition(self):
         return SemEvalPartition(self.train_sents)
 
     def test_partition(self):
-        # TODO use dev set
-        return SemEvalPartition(self.train_sents)
-
-
+        return SemEvalPartition(self.dev_sents)
